# [S3 Select] Eliminate message metadata overhead in SQL Select responses