<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->

This change collects SQL query result records into a buffer (1MiB) and then sends out the data as a single message in the response. This greatly reduces the number of bytes transferred for queries that return small records.

Previously we were sending 1 record per message in the response. Each message has 16 bytes of metadata overhead. This change sends as many (whole) records as needed to fill a 1MiB buffer in a single message.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

The performance gain is dramatic for queries that return small records - take `mys3/sqlselectapi/big.csv.gz` that contains looks like:

```shell
$ mc cat mys3/sqlselectapi/big.csv.gz | head
Id,FirstName,LastName,Rating
1,James,Cruz,3.553518234841988
2,James,Karev,4.500141413056121
3,Alice,Merkel,0.7988159563374936
4,Alice,Karev,1.3823139588325635
5,Angela,Ford,1.4022646931870353
6,Alice,Cruz,3.8760075447547266
7,Joe,Jones,1.0697657081273237
8,Joe,Jones,4.0405880522258455
9,Joe,Merkel,0.1369984718177114
```

It contains 1 million such short rows.

With this file, the following queries were run using mc and measurements made (mc and minio are running on a laptop in this micro-benchmark):

|Query|Master|This PR|
|:------|:------|:-----|
| `select * from s3object` | 15s | 3s |
| `select s.Id from s3object s` | 21s | 4s |
| `select  s.LastName, s.FirstName, s.Rating, s.Id from s3object s ` | 37s | 9s |

All measurements are a fastest of 5 runs rounded to a whole seconds.

Improvements here are 4-5X!

## Regression
<!-- Is this PR fixing a regression? (Yes / No) -->
<!-- If Yes, optionally please include minio version or commit id or PR# that caused this regression, if you have these details. -->

No

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

The results with master and this PR on the above queries were compared for byte-to-byte exactness by hashing the mc output with md5sum.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [x] Performance enhancement

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added unit tests to cover my changes.
- [ ] I have added/updated functional tests in [mint](https://github.com/minio/mint). (If yes, add `mint` PR # here: )
- [x] All new and existing tests passed.