@donatello 2019-01-26
Please note that this PR depends on https://github.com/minio/minio/pull/7102 and can only be merged after that PR.